+++ packages/seo-plugin/src/seo-interface/analysis/types.ts
